pyproject.toml
[project]
name = "restaurant-name-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.13.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langsmith>=0.7.23",
    "pandas>=2.3.3",
    "pinecone>=8.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
]
